Refactoring, moved prompt interfaces into own file in models dir
